requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0
PyMySQL==1.1.0 